writing pubkey at buildtime works
